+++ packages/pwa-kit-dev/src/configs/webpack/config.js
@@ -8,20 +8,26 @@
 /* eslint-env node */
 
 // For more information on these settings, see https://webpack.js.org/configuration
-import path, {resolve} from 'path'
-import fse from 'fs-extra'
-
-import webpack from 'webpack'
-import WebpackNotifierPlugin from 'webpack-notifier'
-import CopyPlugin from 'copy-webpack-plugin'
 import {BundleAnalyzerPlugin} from 'webpack-bundle-analyzer'
+import {resolve} from 'path'
+import fse from 'fs-extra'
+import webpack from 'webpack'
+
+// Third-Party Plugins
+import CopyPlugin from 'copy-webpack-plugin'
 import LoadablePlugin from '@loadable/webpack-plugin'
 import ReactRefreshWebpackPlugin from '@pmmmwh/react-refresh-webpack-plugin'
 import SpeedMeasurePlugin from 'speed-measure-webpack-plugin'
+import WebpackNotifierPlugin from 'webpack-notifier'
 
-import OverridesResolverPlugin from './overrides-plugin'
+// Local Plugins
+import OverridesResolverPlugin from './plugins/overrides-resolver-plugin'
 import {sdkReplacementPlugin} from './plugins'
+
+// Constants
 import {CLIENT, SERVER, CLIENT_OPTIONAL, SSR, REQUEST_PROCESSOR} from './config-names'
+
+// Utilities
 import {getConfig} from '@salesforce/pwa-kit-runtime/utils/ssr-config'
 import {buildAliases, getExtensionNames, nameRegex} from '../../utils/extensibility-utils'
 
@@ -39,24 +45,14 @@ const INSPECT = process.execArgv.some((arg) => /^--inspect(?:-brk)?(?:$|=)/.test
 const DEBUG = mode !== production && process.env.DEBUG === 'true'
 const CI = process.env.CI
 const disableHMR = process.env.HMR === 'false'
+
 const {app: appConfig} = getConfig()
 
 if ([production, development].indexOf(mode) < 0) {
     throw new Error(`Invalid mode "${mode}"`)
 }
 
-// for API convenience, add the leading slash if missing
-export const EXT_OVERRIDES_DIR =
-    typeof pkg?.ccExtensibility?.overridesDir === 'string' &&
-    !pkg?.ccExtensibility?.overridesDir?.match(/(^\/|^\\)/)
-        ? '/' + pkg?.ccExtensibility?.overridesDir?.replace(/\\/g, '/')
-        : pkg?.ccExtensibility?.overridesDir
-        ? pkg?.ccExtensibility?.overridesDir?.replace(/\\/g, '/')
-        : ''
-export const EXT_OVERRIDES_DIR_NO_SLASH = EXT_OVERRIDES_DIR?.replace(/^\//, '')
-export const EXT_EXTENDS = pkg?.ccExtensibility?.extends
-export const EXT_EXTENDS_WIN = pkg?.ccExtensibility?.extends?.replace('/', '\\')
-export const EXT_EXTENDABLE = pkg?.ccExtensibility?.extendable
+export const SUPPORTED_FILE_EXTENSIONS = ['.ts', '.tsx', '.js', '.jsx', '.json']
 
 // TODO: can these be handled in package.json as peerDependencies?
 // https://salesforce-internal.slack.com/archives/C0DKK1FJS/p1672939909212589
@@ -82,18 +78,6 @@ export const DEPS_TO_DEDUPE = [
     '@emotion/react'
 ]
 
-if (EXT_EXTENDABLE && EXT_EXTENDS) {
-    const extendsAsArr = Array.isArray(EXT_EXTENDS) ? EXT_EXTENDS : [EXT_EXTENDS]
-    const conflicts = extendsAsArr.filter((x) => EXT_EXTENDABLE?.includes(x))
-    if (conflicts?.length) {
-        throw new Error(
-            `Dependencies in 'extendable' and 'extends' cannot overlap, fix these: ${conflicts.join(
-                ', '
-            )}"`
-        )
-    }
-}
-
 const getBundleAnalyzerPlugin = (name = 'report', pluginOptions) =>
     new BundleAnalyzerPlugin({
         analyzerMode: 'static',
@@ -107,22 +91,18 @@ const getBundleAnalyzerPlugin = (name = 'report', pluginOptions) =>
     })
 
 const entryPointExists = (segments) => {
-    for (let ext of ['.js', '.jsx', '.ts', '.tsx']) {
-        const primary = resolve(projectDir, ...segments) + ext
-        const override = EXT_OVERRIDES_DIR
-            ? resolve(projectDir, EXT_OVERRIDES_DIR_NO_SLASH, ...segments) + ext
-            : null
-
-        if (fse.existsSync(primary) || (override && fse.existsSync(override))) {
+    for (let ext of SUPPORTED_FILE_EXTENSIONS) {
+        const p = resolve(projectDir, ...segments) + ext
+        if (fse.existsSync(p)) {
             return true
         }
     }
     return false
 }
 
-const getAppEntryPoint = () => {
-    return resolve('./', EXT_OVERRIDES_DIR_NO_SLASH, 'app', 'main')
-}
+const getAppEntryPoint = () => './app/main'
+
+const getServerEntryPoint = () => './app/ssr.js'
 
 const getPublicPathEntryPoint = () => {
     return resolve(
@@ -199,19 +179,18 @@ const baseConfig = (target) => {
                     path: buildDir
                 },
                 resolve: {
-                    ...(EXT_EXTENDS && EXT_OVERRIDES_DIR
-                        ? {
-                              plugins: [
-                                  new OverridesResolverPlugin({
-                                      extends: [EXT_EXTENDS],
-                                      overridesDir: EXT_OVERRIDES_DIR,
-                                      projectDir: process.cwd()
-                                  })
-                              ]
-                          }
-                        : {}),
-                    extensions: ['.ts', '.tsx', '.js', '.jsx', '.json'],
+                    plugins: [
+                        new OverridesResolverPlugin({
+                            projectDir: process.cwd(),
+                            extensions: appConfig?.extensions,
+                            fileExtensions: SUPPORTED_FILE_EXTENSIONS
+                        })
+                    ],
+                    extensions: SUPPORTED_FILE_EXTENSIONS,
                     alias: {
+                        // TODO: This alias is temporary. When we investigate turning the retail template into an application extension
+                        // we'll have to decide if we want to continue using an alias, or change back to using relative paths.
+                        '@salesforce/retail-react-app': projectDir,
                         // Create alias's for "all" extensions, enabled or disabled, as they as they are being imported from the SDK package
                         // and cannot be resolved from that location. We create alias's for all because we do not know which extensions
                         // are configured at build time.
@@ -224,26 +203,7 @@ const baseConfig = (target) => {
                             ...DEPS_TO_DEDUPE.map((dep) => ({
                                 [dep]: findDepInStack(dep)
                             }))
-                        ),
-                        ...(EXT_OVERRIDES_DIR && EXT_EXTENDS
-                            ? Object.assign(
-                                  // NOTE: when an array of `extends` dirs are accepted, don't coerce here
-                                  ...[EXT_EXTENDS].map((extendTarget) => ({
-                                      [extendTarget]: path.resolve(
-                                          projectDir,
-                                          'node_modules',
-                                          ...extendTarget.split('/')
-                                      )
-                                  }))
-                              )
-                            : {}),
-                        ...(EXT_EXTENDABLE
-                            ? Object.assign(
-                                  ...[EXT_EXTENDABLE].map((item) => ({
-                                      [item]: path.resolve(projectDir)
-                                  }))
-                              )
-                            : {})
+                        )
                     },
                     ...(target === 'web' ? {fallback: {crypto: false}} : {})
                 },
@@ -290,7 +250,7 @@ const baseConfig = (target) => {
                             }
                         },
                         {
-                            test: /universal\/extensibility\/extensions/,
+                            test: /universal[\\/]+extensibility[\\/]+extensions/,
                             loader: `@salesforce/pwa-kit-dev/configs/webpack/loaders/extensions-loader`,
                             options: {
                                 pkg
@@ -328,27 +288,7 @@ const withChunking = (config) => {
             splitChunks: {
                 cacheGroups: {
                     vendor: {
-                        // Three scenarios that we'd like to chunk vendor.js:
-                        // 1. The package is in node_modules
-                        // 2. The package is one of the monorepo packages.
-                        //    This is for local development to ensure the bundle
-                        //    composition is the same as a production build
-                        // 3. If extending another template, don't include the
-                        //    baseline route files in vendor.js
-                        test: (module) => {
-                            if (
-                                EXT_EXTENDS &&
-                                EXT_OVERRIDES_DIR &&
-                                module?.context?.includes(
-                                    `${path.sep}${
-                                        path.sep === '/' ? EXT_EXTENDS : EXT_EXTENDS_WIN
-                                    }${path.sep}`
-                                )
-                            ) {
-                                return false
-                            }
-                            return module?.context?.match?.(/(node_modules)|(packages\/(.*)dist)/)
-                        },
+                        test: /(node_modules)|(packages\/.*\/dist)/,
                         name: 'vendor',
                         chunks: 'all'
                     }
@@ -359,31 +299,17 @@ const withChunking = (config) => {
 }
 
 const staticFolderCopyPlugin = new CopyPlugin({
-    patterns: [
-        {
-            from: path
-                .resolve(`${EXT_OVERRIDES_DIR ? EXT_OVERRIDES_DIR_NO_SLASH + '/' : ''}app/static`)
-                .replace(/\\/g, '/'),
-            to: `static/`,
-            noErrorOnMissing: true
-        }
-    ]
+    patterns: [{from: 'app/static/', to: 'static/'}]
 })
 
 const ruleForBabelLoader = (babelPlugins) => {
     return {
         id: 'babel-loader',
         test: /(\.js(x?)|\.ts(x?))$/,
-        ...(EXT_OVERRIDES_DIR && EXT_EXTENDS
-            ? // TODO: handle for array here when that's supported
-              {
-                  exclude: new RegExp(
-                      `${path.sep}node_modules(?!${path.sep}${
-                          path.sep === '/' ? EXT_EXTENDS : EXT_EXTENDS_WIN
-                      })`
-                  )
-              }
-            : {exclude: /node_modules/}),
+        // NOTE: Because our extensions are just folders containing source code, we need to ensure that the babel-loader processes them.
+        // By default babel doesn't process files in "node_modules" folder, so here we will ensure they are included.
+        // TODO: Make sure this regex works for windows.
+        exclude: /^\/node_modules\/(?:@([^/]+)\/)*(?!extension-)[^/]+$/i,
         use: [
             {
                 loader: findDepInStack('babel-loader'),
@@ -479,7 +405,7 @@ const clientOptional = baseConfig('web')
             ...config,
             name: CLIENT_OPTIONAL,
             entry: {
-                ...optional('loader', resolve(projectDir, EXT_OVERRIDES_DIR, 'app', 'loader.js')),
+                ...optional('loader', resolve(projectDir, 'app', 'loader.js')),
                 ...optional('worker', resolve(projectDir, 'worker', 'main.js')),
                 ...optional('core-polyfill', resolve(projectDir, 'node_modules', 'core-js')),
                 ...optional('fetch-polyfill', resolve(projectDir, 'node_modules', 'whatwg-fetch'))
@@ -542,7 +468,7 @@ const ssr = (() => {
                         : {}),
                     // Must *not* be named "server". See - https://www.npmjs.com/package/webpack-hot-server-middleware#usage
                     name: SSR,
-                    entry: `.${EXT_OVERRIDES_DIR}/app/ssr.js`,
+                    entry: getServerEntryPoint(),
                     output: {
                         path: buildDir,
                         filename: 'ssr.js',
@@ -568,8 +494,7 @@ const requestProcessor =
             return {
                 ...config,
                 name: REQUEST_PROCESSOR,
-                // entry: './app/request-processor.js',
-                entry: `.${EXT_OVERRIDES_DIR}/app/request-processor.js`,
+                entry: './app/request-processor.js',
                 output: {
                     path: buildDir,
                     filename: 'request-processor.js',