+++ packages/pwa-kit-dev/src/utils/resolver-utils.js
@@ -10,10 +10,16 @@ import path from 'path'
 const EXTENSION_NAMESPACE = '@salesforce'
 const EXTENSION_PREFIX = 'extension'
 const NODE_MODULES = 'node_modules'
-const SDK_COMPONENT_MAP = {
-    'app/routes': '/ssr/universal/components/routes'
-}
-const INDEX_FILE = 'index'
+const OVERRIDES = 'overrides'
+const APP = 'app'
+const SRC = 'src'
+const PWA_KIT_REACT_SDK = 'pwa-kit-react-sdk'
+
+// TODO: We should determine if we want the `overrides-resolver-plugin` to handle resolution of application special
+// components like _app and _document. If so we can update this map and remove the special logic from our webpack
+// configuration.
+const SDK_COMPONENT_MAP = {}
+const INDEX_FILE = 'index' // TODO: Make this value obey the webpack's `resolve.mainFiles` options.
 
 // Returns true/false indicating if the importPath resolves to a same named file as the sourcePath.
 // @private
@@ -80,44 +86,43 @@ export const expand = (extensions = []) =>
  * @param {String} sourcePath - The path the the file of the source import.
  * @param {Object} opts - The path the the file of the source import.
  * @param {Array<shortName: String, config: Array>} opts.extensions - List of extensions used by the base PWA-Kit application.
- * @param {String>} opts.productDir - Absolute path of the base project.
+ * @param {String>} opts.projectDir - Absolute path of the base project.
  * @returns {String[]} paths - The potential paths to find the module import.
  */
 export const buildCandidatePaths = (importPath, sourcePath, opts = {}) => {
     // Replace wildcard character as it has done its job getting us to this point.
     importPath = importPath.replace('*/', '')
 
-    const {extensions = [], productDir = process.cwd()} = opts
+    const {extensions = [], projectDir = process.cwd()} = opts
     const isSelfReferenceImport = isSelfReference(importPath, sourcePath)
+
     let paths = expand(extensions).reverse()
 
     // Map all the extensions and resolve the module names to absolute paths.
-    paths = paths.map((extension) => {
+    paths = paths.reduce((acc, extensionEntry) => {
         // The reference can be a module/package or an absolute path to a file.
-        const [extensionRef] = extension
-        const isLocalExtension = extensionRef.startsWith(path.sep)
-
-        return path.join(
-            ...(isLocalExtension
-                ? [extensionRef, importPath]
-                : [productDir, NODE_MODULES, extensionRef, importPath])
-        )
-    })
+        const [extensionRef] = extensionEntry
+        const srcPath = path.join(projectDir, NODE_MODULES, extensionRef, SRC)
+        return [...acc, path.join(srcPath, OVERRIDES, importPath), path.join(srcPath, importPath)]
+    }, [])
 
     // Add non-extension search locations locations. The base project and the sdk as the final callback.
     paths = [
         // Base Project
-        path.join(productDir, importPath),
+        path.join(projectDir, APP, OVERRIDES, importPath),
+        path.join(projectDir, APP, importPath),
         // Extensions
         ...paths,
         // SDK
-        path.join(
-            productDir,
-            NODE_MODULES,
-            EXTENSION_NAMESPACE,
-            'pwa-kit-react-sdk',
-            SDK_COMPONENT_MAP[importPath]
-        )
+        ...(SDK_COMPONENT_MAP[importPath]
+            ? path.join(
+                  projectDir,
+                  NODE_MODULES,
+                  EXTENSION_NAMESPACE,
+                  PWA_KIT_REACT_SDK,
+                  SDK_COMPONENT_MAP[importPath]
+              )
+            : [])
     ]
 
     // Under certain circumstances we want to truncate the cadidate path array to prevent circular dependancies.
@@ -127,7 +132,7 @@ export const buildCandidatePaths = (importPath, sourcePath, opts = {}) => {
         // NOTE: Overriding files requires that you use the exact file name, you cannot replace a non-index file with one that
         // is an index file.
         const index = paths.indexOf(sourcePath.split('.')[0])
-        paths = paths.slice(index + 1)
+        paths = paths.slice(index + 2)
     }
 
     return paths