+++ packages/pwa-kit-dev/CHANGELOG.md
@@ -1,4 +1,5 @@
 ## v4.0.0-dev (Jun 21, 2024)
+- Create new `OverrideResolverPlugin` and integrate into the Webpack configuration. [#2012](https://github.com/SalesforceCommerceCloud/pwa-kit/pull/2012)
 - Replace `IApplicationExtension` with `ApplicationExtension` abstract class. [#2019](https://github.com/SalesforceCommerceCloud/pwa-kit/pull/2019)
 - Change Webpack logic to include all installed Application Extensions in the app bundle. [#2004](https://github.com/SalesforceCommerceCloud/pwa-kit/pull/2004)
 - Support the tuples way of configuring the app extensions [#2010](https://github.com/SalesforceCommerceCloud/pwa-kit/pull/2010)